Add dref sorted file for actions
